refactor: Resolve WPS232 reduce module complexity
