+++ cli/core/price_lists/__init__.py
@@ -1,3 +1,3 @@
-from cli.core.price_lists.app import app
+from cli.core.price_lists.app.app import app
 
 __all__ = ["app"]
